refactor(shell): Phase 3a——剪贴板捕获桥外置（含 #ifdef 条件编译验证）
clipboardBridge.js（230 行，4 个方法）：桌面端 Electron 推送 + H5 三路兜底
监听、TEXT/IMAGE/FILE 三类载荷入库、window 级去重。

project-overview.vue 5459 → 5230 行。

这组之前被 Phase 2 刻意跳过，因为含 4 处 #ifdef H5，而当时 .js 模块里
没有条件编译先例，不确定 uni 预处理器是否处理 src/pages/**.js。

先用一次性探针实测确认了机制（探针已回滚，不入库）：
- .vue 内 #ifdef MP-WEIXIN 包裹的标记：产物中消失（已知good基线）
- .js 内同样包裹的标记：产物中同样消失 → 预处理器对 .js 生效
- 两侧未包裹的对照标记：均存在 → 排除被摇树/压缩吃掉的误判

搬移后再反向验证：#ifdef H5 内的代码（已捕获图片/已捕获文件/
__checkbaClipLastText）在 h5 产物中均存在，未被误剥离。

验证：
- 逐行比对：230 行字节一致
- 方法名扫描：七个模块 62 个方法与 .vue 剩余 152 个方法零重名
- npm run build:h5 通过；npm run check:emits 通过
- npm run test:app-e2e 连续 3 次 29/29

注：app-e2e 不覆盖剪贴板链路，本次信心来自静态等价证明而非端到端。
